Fix cyclic dependency by joining parameter.ts with transformation.ts
